requirements.txt
Flask==2.3.3
requests==2.31.0
python-dotenv==1.0.0
openai==0.28.1
flask-cors==4.0.0
